+++ lib/src/lexeme.dart
@@ -1,7 +1,4 @@
 import 'package:token_parser/src/error.dart';
-import 'package:token_parser/src/grammar.dart';
-import 'package:token_parser/src/lexemes/until.dart';
-import 'package:token_parser/src/token.dart';
 
 /* -= Lexeme Imports =- */
 
@@ -22,10 +19,12 @@ import 'package:token_parser/src/lexemes/start.dart';
 import 'package:token_parser/src/lexemes/end.dart';
 
 import 'package:token_parser/src/lexemes/character.dart';
+import 'package:token_parser/src/lexemes/until.dart';
 import 'package:token_parser/src/lexemes/repeat.dart';
 
 import 'package:token_parser/src/lexemes/reference.dart';
 import 'package:token_parser/src/lexemes/self.dart';
+import 'package:token_parser/token_parser.dart';
 
 /* -=-=-=-=-=-=-=-=-=- */
 
@@ -195,7 +194,9 @@ abstract class Lexeme extends Pattern {
   factory Lexeme.character(Pattern pattern, { String? name }) = CharacterLexeme;
   factory Lexeme.any() = CharacterLexeme.any;
   factory Lexeme.until(Pattern pattern, Pattern until, { String? name }) = UntilLexeme;
+  
   factory Lexeme.repeat(Pattern pattern, int min, { int? max, String? name }) = RepeatLexeme;
+  factory Lexeme.global(Pattern pattern) => (pattern | any()).multiple;
 
   factory Lexeme.reference(String name, { Grammar? grammar }) = ReferenceLexeme;
   factory Lexeme.ref(String name, { Grammar? grammar }) = ReferenceLexeme;